requirements.txt
astroid==3.3.10
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cx_Oracle==8.3.0
cycler==0.12.1
dill==0.4.0
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
-e git+https://github.com/milaptce/monitoramento_oracle.git@e26395928461b2ecc9dc9c286ddcf466f4e0614e#egg=monitor_oracle_fts
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
py-cpuinfo==9.0.0
Pygments==2.19.1
pylint==3.3.7
pyparsing==3.2.3
pytest==8.4.0
pytest-benchmark==5.1.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
six==1.17.0
smmap==5.0.2
sql_metadata==2.17.0
sqlparse==0.5.3
tomlkit==0.13.3
tzdata==2025.2
urllib3==2.4.0
